pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.16.5",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.128.0",
    "fastapi-login>=1.10.3",
    "greenlet>=3.2.4",
    "passlib>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-to-typescript>=2.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.39.0",
]
